add xkcd command + refactor

+++ src/commands/prune.js
@@ -14,18 +14,18 @@ module.exports = {
     if (parseInt(args[0]) == 100 && limit == 101)
       limit--;
 
-    if (!limit) throw "ArgError: limit |is undefined";
+    if (!limit || limit == undefined) throw "ArgError: limit |is undefined";
     if (limit > 100) throw "ArgError: limit |should be less than or equal to 100";
 
     message.channel.fetchMessages({
       limit: limit
     }).then(messages => {
       messages.forEach((msg) => {
-        if (msg.channel.type == 'text') queue.push(msg); // safe guard, probably wastes memory
+        if (msg.channel.type == 'text') queue.push(msg);
       });
 
       message.channel.bulkDelete(queue, false).then(messages => {
-        message.channel.send(`Pruned ${args[0]} messages successfully`)
+        message.channel.send(`Purged ${args[0]} messages successfully`);
       });
     }).catch(console.error);
   },
